Move create_apk_elf_path() into symbolize module
With the rework of the normalization APIs a while back, the
create_apk_elf_path() function is no longer used by the normalization
code itself -- yet it is defined in there.
Move it over into the symbolize module, where it is actually used.
