+++ src/normalize/mod.rs
@@ -55,6 +55,5 @@ pub use normalizer::Normalizer;
 pub use normalizer::Output;
 pub use user::UserOutput;
 
-pub(crate) use user::create_apk_elf_path;
 pub(crate) use user::normalize_sorted_user_addrs_with_entries;
 pub(crate) use user::Handler;
